Add Travis CI configuration
Fix environment variables

The build is failing because the environment variables file was not
found on the build server. Since the actual variables are set in the
Travis CI environment and not a file, it should fix the problem to copy
the example environment variables file and allow empty values.