Remove stdout logging stream
The reason for the low test coverage in the log service is the
conditional that is not executed in the test environment. All this
conditional did was add a log output for `stdout`, which is functionally
useless because it outputs JSON to the console. Removing the `stdout`
stream serves to both increase test coverage and remove an annoyance
when running the application.