Add authentication to all routes
Now, a request to any route except `/` must have the header
`Authorization` with the value `Bearer ` followed by a JWT. The JWT must
have a payload including `sub`, where `sub` (subject) is the ID of an
organization. This will be decoded in each request in order to help the
next handler in the endpoint handler chain which organization to look
for (this is important because data for all organizations is stored in
the same tables).