Add function to build "where" clauses for queries
This allows database service functions to include a check for the
organization ID without having to write repetitive queries. The
`organizationID` parameter is optional because it is obviously the
primary key for the `organization` table, so it would be redundant to
include it twice in queries for that table. Also, it seems overly
prescriptive to require an `organizationID` in *every* future database
table.