Fix password validation bug and add test for it
This ensures that the password validation function `bcrypt.compare`
actually completes its process before the next step of the promise chain
executes, then the `===` ensures that the value of `valid` is actually
`true`, not just a truthy value. The test ensures that this function
actually returns an error for an invalid combination of username and
password.