Add auth error when organization not found
Now, if an organization is not found when authenticating, an error will
be returned.

This error is `401 Not Authorized` instead of `404 Not Found` because a
`404` would reveal that the username is incorrect, not just the
password. This information could be used to differentiate correct email
+ incorrect password from incorrect email + incorrect password, allowing
attackers to focus their efforts and try login combinations more
efficiently. See http://security.stackexchange.com/a/17820 for further
discussion.

This also fixes #73, because the wrong column name was being used for
`organizationID`.