Change from itemID to tag on item endpoints
This is more convenient for the frontend, which will need to access
items by tag, because that is how the user will scan items in. The
`itemID` field is still the primary key instead of `tag`, however,
because it is a smaller value than the `tag` column and therefore will
index more efficiently and help keep database size small.