Move endpoint definitions into controllers
Instead of defining all routes in `./controllers/routes.js`, define
routes inline with their controllers. This prevents any single file from
becoming bloated as more and more endpoints are added to the
application. It also moves the endpoint definitions closer to their
actual logic, which is useful for reducing the cognitive load of
developing the application.

Each controller has a new `mount()` function which, when called with a
Restify application as its parameter, will mount endpoints on the
application. All controllers still export their functions in the same
way as before, which retains their testability and their ability to be
used as components of other modules.