Remove paramName parameter from endpoint service
This parameter was added with the intention of allowing the database
column to have a different name than the name of the parameter in the
path definition. However, this has not been used yet, and there is no
particular reason that the parameter would have a different name than
the database column.

Since it was optional, this parameter would get in the way of having
a *second* optional parameter for all of the functions in this
module. This new optional parameter will be a function used to add
clauses to the database query. If `paramName` was kept, the functions
would have to be called with a `null` in its place in order to use the
new parameter.