Add overrriding of default messages in endpoint
Now, each endpoint can have a fully custom set of messages for endpoint
actions and errors. This allows error messages to be relevant to each
endpoint without needing to bypass the endpoint service.

Since both the `modify` function and the `messages` object are optional
parameters, this required adding an optional parameters object as the
last parameter of each endpoint handler function in the endpoint
service. Now, both parameters can be optional, but with destructuring,
the parameters are still available as `modify` and `messages`.