Change hash to map to fix for..in problem
When using an object hash as a map, a for..in loop needs to check that
each property is actually defined on the object itself. To avoid that,
an ES6 Map can be used. This has the side effect of allowing the use of
a for..of loop, which is cleaner than the for..in loop.