Fix automatic organization ID extraction bug
Automatically adding the `organizationID` to the request body is a
good feature for most endpoints to have, but causes conflicts on the
organization endpoints when a user tries to create another
organization.

It may be a desired feature to prevent users from creating more than
one organization, but it should be implemented as an intentional
feature instead of an unexpected side effect.