Change get rental endpoint to use rentalID
This endpoint had been using barcode as a parameter, which was not
consistent with the docs. This approach also made use of the implicit
fact that the `barcode` column would be available in the query *at
some point* (in this case via the modify function
`rental.withBarcode`). This wasn't wrong in itself, but must be
removed now that `db.buildWhere` aliases all column names with their
respective table names.