Fix get item barcode parameter
Aliasing the column name with its table name did not work as a solution
to the ambiguity error. It would have worked if the column name had
been passed only to the database, but it was also used as a way to get
the parameter from the request. This in turn caused the query to not
use the parameter from the request (because the names no longer
matched) and caused all rows to be returned. But, since only the
`.first()` row is returned, it looked like the same row was returned
no matter what barcode was passed.