Fix auth failure when refresh token exists
The authenticate endpoint was failing when the refresh token already
existed in the database because of a duplicate entry on the primary
key `userID`. This solves the problem by checking if the token is
already in the database and updating it when it already exists instead
of naively trying to insert it again.