272/add Stripe endpoints (#281)
* Add Stripe

* Add Stripe keys

* Add email and Stripe customer fields to organization

Closes #267.

* Move password hashing into a method

This will allow other controllers to hash passwords without using Bcrypt directly.

* Add subscription endpoint

* Update VS Code tasks

* Fix syntax errors

* Add default test task to VS Code config

* Fix change password

This was broken by the removal of `auth.saltRounds`, but can easily be fixed by replacing the direct use of bcrypt with `auth.hashPassword`.

* Fix environment variables loading too late

* Set first user to admin and include created IDs in response

The first user in the organization should be an administrator so that he or she can add and administer other users. Including the created IDs in the response is helpful for then retrieving those entities.

* Fix promise indentation

* Add subscription tests

This involved removing many unnecessary conditionals and error checks, because they will be unreachable in production usage, and if they are reached, it is a symptom of a bigger problem.

* Fix indentation

* Fix test name capitalization

* Fix cleaning up created Stripe customers

* Remove "create organization" endpoint

This functionality is provided by the new "create subscription" endpoint.