Fix get item custom field missing results
When a custom field applies to all categories (it has no rows in `customFieldCategory`) and has no value, it was not being returned by this endpoint because the `where` statement expects `itemCustomField.barcode` to be set. Allowing that field to be `null` ensures that these custom fields will be present in the response.