Fix cross-organization categories in item custom fields
The query for item custom fields included a `join` with all items that matched on `categoryID`, which would only match with items in the same organization, or `categoryID is null`, which would match with all items in the database. The latter could cause strange problems. Adding an `and organizationID` to the `join` fixes the problem.