Fixing update rental v2 failing
The adapter method that maintains backward compatibility for v2 was always adding the properties `start` and `end` to the request body, even when `startDate` and `endDate` were not present in the request. This was causing the conditional meant to prevent empty updates to not work.

Refactoring the method to provide a no-op update to the rental fixes the problem of the endpoint never returning (because it would not call `res.send` when only the `returnDate` was provided).