Add custom field type
Custom field values are automatically cast to a string when being inserted into the database. Now, with the addition of `customField.fieldTypeID` and the table `fieldType`, custom field values are type cast back to the correct type before sending them to the user.

This implementation uses  the partial hack of overriding `res.send` before calling the endpoint service. A better way to implement it would be to pass the endpoint service a function that modifies results before sending them back to the user, but this is the only controller that has this use case, so it makes more sense to rely on the current implementation for now.